% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adapt_wrapper.R
\name{adapt_mtp}
\alias{adapt_mtp}
\title{Wrapper of AdaPT multiple testing procedure}
\usage{
adapt_mtp(Ps, Xs, alpha, return_fit = FALSE)
}
\arguments{
\item{Ps}{Numeric vector of p-values}

\item{Xs}{Data frame with features}

\item{alpha}{Nominal testing level}

\item{return_fit}{Boolean, whether to return the fitted adapt object (or only the indicator of rejections), defaults to false}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Wrapper of AdaPT multiple testing procedure
}
\references{
AdaptMT CRAN package
}
\seealso{
\url{https://github.com/Huber-group-EMBL/covariate-powered-cross-weighted-multiple-testing/blob/master/IHWStatsPaper/R/adapt_wrapper.R}
}
