% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adapt_wrapper.R
\name{adapt_xgboost_cv_wrapper}
\alias{adapt_xgboost_cv_wrapper}
\title{Wrapper for AdaPT Wrapper devtools::install_github("ryurko/adaptMT")}
\usage{
adapt_xgboost_cv_wrapper(
  Ps,
  Xs,
  alphas = 0.1,
  args_search = list(nrounds100md2 = list(nrounds = 15, max_depth = 3, verbose = 0,
    nthread = 2))
)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Xs}{Vector or matrix of covariates}

\item{alpha}{Significance level at which to apply method

Xs <- runif(20000, min=0, max=2.5) # covariate
Hs <- rbinom(20000,1,0.1) # hypothesis true or false
Zs <- rnorm(20000, Xs*Hs) # Z-score
Ps <- 1-pnorm(Zs) # pvalue
adapt_xgboost_cv_wrapper(Ps, Xs)}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Wrapper for AdaPT Wrapper devtools::install_github("ryurko/adaptMT")
}
