% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adapt_wrapper.R
\name{adapt_xgboost_wrapper}
\alias{adapt_xgboost_wrapper}
\title{Wrapper for AdaPT Wrapper devtools::install_github("ryurko/adaptMT")}
\usage{
adapt_xgboost_wrapper(Ps, Xs, alpha = 0.1, nfits = 5)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Xs}{Vector or matrix of covariates}

\item{alpha}{Significance level at which to apply method}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Wrapper for AdaPT Wrapper devtools::install_github("ryurko/adaptMT")
}
\examples{
Xs <- runif(2000, min=0, max=2.5) # covariate
Hs <- rbinom(2000,1,0.1) # hypothesis true or false
Zs <- rnorm(2000, Xs*Hs) # Z-score
Ps <- 1-pnorm(Zs) # pvalue
adapt_xgboost_wrapper(Ps, Xs)

}
