% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_flexible_rule.R
\name{ball_volume}
\alias{ball_volume}
\title{Calculate the volume of an n-dimensional ball in the p-norm}
\usage{
ball_volume(radius, lp_norm, dimension)
}
\arguments{
\item{radius}{The radius of the ball.}

\item{lp_norm}{The norm to be used. Must be a positive number.}

\item{dimension}{The number of dimensions.}
}
\value{
The volume of the n-dimensional ball.
}
\description{
Calculate the volume of an n-dimensional ball in the p-norm
}
\examples{
ball_volume(1, 2, 3)
ball_volume(2, 2, 3)
ball_volume(1, 1, 2)
}
