% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_boca_leek_wrapper.R
\name{boca_leek_wrapper}
\alias{boca_leek_wrapper}
\title{Boca-Leek Wrapper Function for Prior Probability Estimation}
\usage{
boca_leek_wrapper(
  pvalues,
  covariates,
  alpha,
  nfolds = 5,
  ntrees = 10,
  tau = 0.5,
  nodedepth = NULL,
  nodesize = 300,
  mtry = "auto",
  seed = NULL
)
}
\arguments{
\item{pvalues}{Numeric vector of unadjusted p-values.}

\item{covariates}{Matrix which contains the covariates (independent under the H0 of the p-value) for each test.}

\item{alpha}{Double, significance level for the adjusted p-values.}

\item{nfolds}{Integer, number of folds to use for cross-validation (default is 5).}

\item{ntrees}{Integer, see same parameter in \code{\link[randomForestSRC]{rfsrc}} (default is 10).}

\item{tau}{Double, censoring threshold tau of the p-values in the stratification method "forest" (default is 0.5).}

\item{nodedepth}{Integer, see same parameter in \code{\link[randomForestSRC]{rfsrc}}.}

\item{nodesize}{Integer, see same parameter in \code{\link[randomForestSRC]{rfsrc}} (default is 300).}

\item{mtry}{Integer, see same parameter in \code{\link[randomForestSRC]{rfsrc}}. Use "auto" for automatic selection (default is "auto").}

\item{seed}{Integer, specifies the random seed to be used.}
}
\value{
Logical vector indicating if the adjusted p-values are below the significance level alpha.
}
\description{
This function serves as a wrapper for the `estimate_prior_prob` function,
applying a random forest approach to estimate prior probabilities and adjust p-values
based on the Benjamini-Hochberg method. It then checks if the adjusted p-values are
below the given significance level alpha.
}
