% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_handling.R
\name{create_dataframe}
\alias{create_dataframe}
\title{Expand each vector in a list with the first element of all other vectors}
\usage{
create_dataframe(list_of_vectors)
}
\arguments{
\item{list_of_vectors}{A list of numeric vectors.}
}
\value{
A data.frame where each row is a combination of each vector in the input list
        with the first element of all other vectors.
}
\description{
Expand each vector in a list with the first element of all other vectors
}
\examples{
v1 <- c(1, 2, 3)
v2 <- c(4, 5, 6)
v3 <- c(7, 8, 9)
list_of_vectors <- list(v1, v2, v3)
df <- create_dataframe(list_of_vectors)
print(df)
}
