% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_handling.R
\name{create_formula}
\alias{create_formula}
\title{Create a formula from a data frame}
\usage{
create_formula(Xs)
}
\arguments{
\item{Xs}{A data frame whose column names will be used to create the formula.}
}
\value{
A formula object, with the right-hand side containing all the column names.
}
\description{
This function takes a data frame and constructs a formula
based on its column names.
}
