% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_flexible_rule.R
\name{discrete_prop_alt_creator}
\alias{discrete_prop_alt_creator}
\title{Create a function to generate discrete alternative probabilities.}
\usage{
discrete_prop_alt_creator(
  seed,
  dimensions,
  ndim = 1,
  signal_strength = 0.9,
  lp_norm = 1,
  target_average_alt_prob = 0.1
)
}
\arguments{
\item{seed}{The seed for random number generation (optional).}

\item{dimensions}{The total number of dimensions in the covariate matrix.}

\item{ndim}{The number of dimensions to consider when calculating the alternative probability (default is 1).}

\item{signal_strength}{signal_strength, between 0 and 1}

\item{lp_norm}{The norm to consider for the calculation (default is 1).}

\item{target_average_alt_prob}{The target average for the alternative probability.}
}
\value{
A function that generates discrete alternative probabilities based on input covariate values.
}
\description{
This function creates a function that generates discrete alternative probabilities based on random covariate values.
 The volume of an n-dimensional ball in the p-norm (also known as the L^p norm) is given by the formula:
}
\examples{
# Create an alternative probability generator with 5 dimensions and considering 2 dimensions
alt_prob_generator <- discrete_prop_alt_creator(seed = 42, dimensions = 5, ndim = 2, c = 0.9)
# Generate alternative probability for a covariate vector
alt_prob <- alt_prob_generator(c(0.2, 0.5, 0.1, 0.8, 0.4))

}
