% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_boca_leek_wrapper.R
\name{estimate_prior_prob}
\alias{estimate_prior_prob}
\title{Estimate prior probability with random forest}
\usage{
estimate_prior_prob(
  pvalues,
  covariates,
  folds,
  ntrees = 10,
  tau = 0.5,
  nodedepth = NULL,
  nodesize = 300,
  mtry = "auto",
  seed = NULL
)
}
\arguments{
\item{pvalues}{Numeric vector of unadjusted p-values.}

\item{covariates}{Matrix which contains the covariates (independent under the H0 of the p-value) for each test.}

\item{folds}{Integer vector, Pre-specify assignment of hypotheses into folds.}

\item{ntrees}{Integer, see same parameter in \code{\link[randomForestSRC]{rfsrc}}}

\item{tau}{Double, censoring threshold tau of the pvalues in the stratification method "forest". See more in group_by_forest}

\item{nodedepth}{Integer, see same parameter in \code{\link[randomForestSRC]{rfsrc}}}

\item{nodesize}{Integer, see same parameter in \code{\link[randomForestSRC]{rfsrc}}}

\item{mtry}{Integer, see same parameter in \code{\link[randomForestSRC]{rfsrc}}
Use "auto" for automatic selection.}

\item{seed}{Integer, specifies random seed to be used}
}
\value{
prior probabilites
}
\description{
Hypotheses are stratified into bins based on random forest construction, alternative to \code{prior_prob_by_filter_multivariate}
  prior_prob are homogenous wrt to Storeys null proportion estimator
}
\details{
see httpvalues://doi.org/10.7717/peerj.6035 for details on BocaLeek construction
}
\examples{

save.seed <- .Random.seed
set.seed(1)
X <- runif(20000, min = 0, max = 2.5) # covariate
H <- rbinom(20000, 1, 0.1) # hypothesis true or false
Z <- rnorm(20000, H * X) # Z-score
folds <- sample(1:5, 20000, replace = TRUE)
.Random.seed <- save.seed
pvalue <- 1 - pnorm(Z) # pvalue
prior_prob <- estimate_prior_prob(pvalue, as.matrix(X), folds)
}
