% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{eval_sim_parallel}
\alias{eval_sim_parallel}
\title{Evaluate Simulation Parallelly}
\usage{
eval_sim_parallel(
  simulation_list,
  methods = c("IHW-quantile", "IHW-forest", "IHW-forest-drop-inbag", "BH", "AdaPT",
    "Boca-Leek", "Clfdr-EM"),
  forest_par = NULL,
  null_proportion = TRUE,
  folds = NULL,
  parallel = TRUE,
  silent = FALSE,
  per_covariate_bins = 5
)
}
\arguments{
\item{simulation_list}{A list of simulations, where each simulation is a list containing:
\itemize{
  \item \code{seed}: The seed value used for random number generation.
  \item \code{pvalue}: The p-values used in the simulation.
  \item \code{covariate}: The covariate values used in the simulation.
  \item \code{Hs}: Hypothesis status or any other relevant information required for the analysis.
  \item \code{alpha}: The alpha level used in the analysis.
  \item Other context information with a length of one (optional).
}}

\item{methods}{Vector of methods to be used for multiple testing procedures.}

\item{forest_par}{Parameters for the forest method.}

\item{null_proportion}{Logical, if TRUE the proportion of null hypotheses is estimated.}

\item{folds}{The number of folds for cross-validation.}
}
\value{
A list of results from running the simulations.
}
\description{
This function is designed to evaluate a list of simulations in parallel using a user-specified number of cores.
The simulations are run for each method given in the `methods` argument.
}
\examples{
sim_list <- list(
  list(seed = 12345, pvalue = c(0.05, 0.10, 0.20), covariate = c(1.2, 3.4, 5.6), Hs = c(TRUE, FALSE, TRUE), alpha = 0.05),
  list(seed = 67890, pvalue = c(0.01, 0.03, 0.05), covariate = c(2.3, 4.5, 6.7), Hs = c(FALSE, FALSE, TRUE), alpha = 0.1)
)
results <- eval_sim_parallel(simulation_list = sim_list, context_information = c("info1", "info2"))
}
