% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_handling.R
\name{filter_df}
\alias{filter_df}
\title{Filter a data frame based on a vector of values}
\usage{
filter_df(df, vec)
}
\arguments{
\item{df}{A data frame to be filtered.}

\item{vector}{A named vector where names correspond to columns in df. Each value in the vector represents
the value to filter on for the corresponding column in df.}
}
\value{
A data frame containing only the rows where for each column, the value in that column is equal to
        the corresponding value in the input vector.
}
\description{
This function filters rows of a data frame such that for each column, the value in that column is equal to
the corresponding value in a provided vector. The vector should be named with names corresponding to 
column names in the data frame.
}
\examples{
df <- data.frame(a = 1:5, b = 6:10, c = 11:15)
vector <- c(a = 2, b = 7)
filter_df(df, vector)
}
