% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_flexible_rule.R
\name{flexible_prop_alt_sim}
\alias{flexible_prop_alt_sim}
\title{Simulate alternative proportions flexibly}
\usage{
flexible_prop_alt_sim(
  m,
  r,
  dimensions,
  prop_alt_function_creator = discrete_prop_alt_creator,
  prop_alt_function_name = "discrete_prop_alt_creator",
  additional_arguments_prop_alt_function_creator = NULL,
  kappa = 0
)
}
\arguments{
\item{m}{A vector specifying the number of observations per simulation}

\item{r}{An integer specifying the number of replications for the simulation}

\item{dimensions}{A vector specifying the dimensionality of the covariates in the simulations}

\item{prop_alt_function_creator}{A function that creates alternative proportion functions, default is 'discrete_prop_alt_creator'}

\item{prop_alt_function_name}{A string naming the proportion alternative function creator, default is "discrete_prop_alt_creator"}

\item{additional_arguments_prop_alt_function_creator}{A dataframe that contains additional arguments needed for the prop_alt_function_creator function}
}
\value{
A list containing the simulation results for each combination of m, dimension, and seed, including additional arguments if provided
}
\description{
This function is designed to perform multiple simulations 
to calculate the proportion of alternative cases given various conditions.

The function assumes the following model:
\deqn{
  \begin{aligned}
  &X_i \stackrel{\text{iid.}}{\sim} \operatorname{U}[-0.5,0.5]^d, \\
  &H_i \mid X_i \sim  \operatorname{Bernoulli}(1-\pi_0(\textcolor{red}{X_i}))\\
  &P_i \mid H_i = 0, X_i \stackrel{\text{iid.}}{\sim} (1-\kappa) \operatorname{U}[0,1]+\kappa \operatorname{Beta}(1,0.5),\\
  &P_i \mid H_i = 1, X_i  \sim \operatorname{Beta}(0.25,1).
  \end{aligned}
}
}
\examples{
flexible_prop_alt_sim(m = c(100,200), r = 2, dimensions = c(2,3))

}
