% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_handling.R
\name{get_split}
\alias{get_split}
\title{Split a dataframe into smaller parts and retrieve a specific split by index.}
\usage{
get_split(df, num_splits, split_index)
}
\arguments{
\item{df}{A dataframe that needs to be split.}

\item{num_splits}{The number of splits you want to divide the dataframe into.}

\item{split_index}{The index of the split you want to retrieve.}
}
\value{
A dataframe corresponding to the specified split index.
}
\description{
This function divides a dataframe (`df`) into a specified number of smaller splits 
and returns one of the splits based on the provided index.
}
\examples{
\dontrun{
data <- data.frame(a = rnorm(100), b = rnorm(100))
result_df <- get_split(data, num_splits = 10, split_index = 3)
}

}
