% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_high_dim.R
\name{high_dim_sim}
\alias{high_dim_sim}
\title{High Dimensional Simulation Function}
\usage{
high_dim_sim(m, r, dimensions)
}
\arguments{
\item{m}{Integer. Defines the number of simulations to perform.}

\item{r}{Integer. Defines the number of unique random seeds to use.}

\item{dimensions}{Integer vector. Each element specifies a number of dimensions to simulate.}
}
\value{
A list of lists. Each sublist contains simulation results for each seed 
and dimension combination, including the covariate matrix, proportion of alternative
hypothesis cases, boolean vector of hypothesis tests, p-values for each simulation,
dimension of the simulation, the seed used, and the number of simulations.
}
\description{
This function creates simulations in a high dimensional space. The dimensions,
random seed, and number of simulations are user-defined. Simulations are 
created using Bernoulli, Exponential distributions and specific beta values.
This function can be useful for simulations in high dimensional analysis.
}
\details{
The key equations for the simulation are:
\deqn{\beta_{\pi} = [3, 3, \underbrace{0, \dots, 0}_{d-2}]}
\deqn{\beta_{\mu} = [2, 2, \underbrace{0, \dots, 0}_{d-2}]}
\deqn{\beta_{0_{\mu}} = 0}

\deqn{X \sim U[0,1]^{d}}
\deqn{\pi_1(X) = \left(\frac{1}{{1 + e^{-(X \times \beta_{\pi} + \beta_{0_{\pi}})}}}\right)}
\deqn{\mu(X) = \max(1, X \times \beta_{\mu} + \beta_{0_{\mu}})}
\deqn{H \mid X \sim \text{{Bernoulli}}(\pi(X))}
\deqn{Y \mid H = 1 \sim \text{{Exponential}}(1/\mu(X))}
\deqn{Y \mid H = 0 \sim \text{{Exponential}}(1)}
\deqn{P = e^{-Y}}
}
\examples{
high_dim_sim(1000, 5, c(2, 4, 6))

}
