% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_ihw_wrapper.R
\name{ihw_quantile_wrapper}
\alias{ihw_quantile_wrapper}
\title{Wrapper for the IHW-Quantile procedure}
\usage{
ihw_quantile_wrapper(
  Ps,
  Xs,
  alpha,
  per_covariate_bins = 5,
  null_proportion = T,
  timeout = 300,
  folds = NULL
)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Xs}{Vector or matrix of covariates.}

\item{alpha}{Significance level at which to apply method.}

\item{per_covariate_bins}{Number of bins for the covariate stratification. Default is 5.}

\item{null_proportion}{Logical; if TRUE, estimates the null proportion. Default is TRUE.}

\item{timeout}{Maximum amount of time (in seconds) for the procedure to run. Default is 300.}

\item{folds}{The number of folds for cross-validation. Default is NULL.}
}
\value{
Binary vector of rejected/non-rejected hypotheses. If the procedure fails or times out, returns NA.
}
\description{
This function applies the IHW-Quantile procedure to a given set of p-values and covariates.
}
\examples{
# Generate some example data
Ps <- runif(100)  # P-values
Xs <- rnorm(100)  # Covariates
# Run the IHW-Quantile procedure
rejections <- ihw_quantile_wrapper(Ps, Xs, alpha = 0.05)
}
