% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_flexible_rule.R
\name{in_ball}
\alias{in_ball}
\title{Check if a vector is within a ball of a given radius in the p-norm}
\usage{
in_ball(vector, radius, lp_norm)
}
\arguments{
\item{vector}{A numeric vector.}

\item{radius}{The radius of the ball.}

\item{lp_norm}{The norm to be used. Must be a positive number.}
}
\value{
TRUE if the vector is within the ball, FALSE otherwise.
}
\description{
Check if a vector is within a ball of a given radius in the p-norm
}
\examples{
in_ball(c(1, 2, 3), 4, 2)
in_ball(c(1, 2, 3), 2, 2)
in_ball(c(1, 2, 3), 1, 1)
}
