% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_handling.R
\name{latest_files}
\alias{latest_files}
\title{Extract and Combine Latest Files Based on File Ending}
\usage{
latest_files(directory, file_ending)
}
\arguments{
\item{directory}{A character string specifying the path of the directory where the files are located.}

\item{file_ending}{A character string specifying the file ending to look for in the directory.}
}
\value{
A combined data frame of all the files with the latest date and the specific file ending in the directory.
}
\description{
This function identifies and combines the latest files in a specific directory based on file ending.
}
\examples{
# For "_eval_high_dim_sim.Rds"
high_dim_res_sim <- latest_files("simulation/data", "_eval_high_dim_sim.Rds")
# For "_eval_high_dim_sim_adapt.Rds"
high_dim_res_adapt <- latest_files("simulation/data", "_eval_high_dim_sim_adapt.Rds")
}
