% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_noise.R
\name{noise_sim}
\alias{noise_sim}
\title{Noise Simulation Function}
\usage{
noise_sim(m, r, dimensions)
}
\arguments{
\item{m}{Integer. Defines the number of simulations to perform.}

\item{r}{Integer. Defines the number of unique random seeds to use.}

\item{dimensions}{Integer vector. Each element specifies a number of dimensions to simulate.}
}
\value{
A list of lists. Each sublist contains simulation results for each seed 
and dimension combination, including the covariate matrix, proportion of alternative
hypothesis cases, boolean vector of hypothesis tests, p-values for each simulation,
dimension of the simulation, the seed used, and the number of simulations.
}
\description{
This function generates noise simulations in a high dimensional space. 
The dimensions, random seed, and number of simulations are user-defined.
This function can be useful for simulations in high dimensional analysis,
especially when the impact of noise needs to be considered.
}
\examples{
noise_sim(1000, 5, c(2, 4, 6))

}
