% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_2d_simulation}
\alias{plot_2d_simulation}
\title{Plot a 2D Simulation}
\usage{
plot_2d_simulation(Xs, Hs, title_list = NULL)
}
\arguments{
\item{Xs}{A matrix or data frame containing the 2D coordinates. It must have two columns.}

\item{Hs}{A vector containing the categories or labels for the points. It must have the same length as the number of rows in Xs.}

\item{title_list}{An optional named list containing the title information. If provided, the names and elements will be concatenated and used as the title for the plot.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function takes two-dimensional data (Xs) and a corresponding vector of categories (Hs) to create a scatter plot. An optional title can be added using a named list with multiple elements.
}
\examples{
Xs <- matrix(rnorm(100), ncol = 2)
Hs <- sample(letters[1:2], 50, replace = TRUE)
title_list <- list(main = "My 2D Simulation", subtitle = "An example plot")
plot_2d_simulation(Xs, Hs, title_list)
}
