% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fdr_power}
\alias{plot_fdr_power}
\title{Plot FDR and Power}
\usage{
plot_fdr_power(
  sim_res,
  group_by_dimension = "length",
  alpha = 0.1,
  log_trans = TRUE
)
}
\arguments{
\item{sim_res}{A data frame containing simulation results.}

\item{group_by_dimension}{The column name in \code{sim_res} to be used for grouping.}

\item{alpha}{The significance level used to plot the FDR threshold line.}
}
\value{
A combined ggplot object showing FDR and Power plots with a common legend.
}
\description{
This function takes a data frame \code{sim_res} with simulation results and generates
two plots showing the False Discovery Rate (FDR) and Power for different grouping 
dimensions and methods.
}
\examples{
sim_res <- data.frame(
  length = rep(c(10, 20, 30), each = 5),
  method = rep(c("A", "B"), each = 15),
  FDP = c(0.1, 0.2, 0.05, 0.15, 0.03, 0.08, 0.25, 0.18, 0.09, 0.12,
          0.2, 0.15, 0.07, 0.05, 0.1, 0.2, 0.25, 0.3, 0.35, 0.4),
  pow = c(0.9, 0.8, 0.85, 0.92, 0.88, 0.92, 0.81, 0.85, 0.9, 0.95,
          0.7, 0.75, 0.8, 0.82, 0.78, 0.72, 0.68, 0.65, 0.62, 0.6)
)
plot_fdr_power(sim_res, group_by_dimension = "length")

}
