% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{run_sim}
\alias{run_sim}
\title{Simulates and evaluates multiple testing procedures.}
\usage{
run_sim(
  Ps,
  Xs,
  Hs = NULL,
  seed = NULL,
  alpha = 0.1,
  forest_par = NULL,
  null_proportion = T,
  methods = c("IHW-quantile", "IHW-forest", "BH", "AdaPT", "Boca-Leek", "Clfdr-EM"),
  folds = NULL
)
}
\arguments{
\item{Ps}{Vector of p-values.}

\item{Xs}{Vector of covariates.}

\item{Hs}{Vector with indicators of alternatives (1) and true nulls (0). Default is NULL.}

\item{seed}{Seed for reproducibility.}

\item{alpha}{Significance level. Default is 0.1.}

\item{forest_par}{Parameter for the forest method.}

\item{null_proportion}{Logical; if TRUE, estimates the null proportion. Default is TRUE.}

\item{methods}{Vector of methods to run. Default includes all valid methods: "IHW-quantile", "IHW-forest", "IHW-forest-drop-inbag", "BH", "AdaPT", "Boca-Leek", "Clfdr-EM", "AdaPT-xgboost".}
}
\value{
Data frame with columns for each method and associated statistics.
}
\description{
Simulates and evaluates multiple testing procedures.
}
\examples{
save.seed <- .Random.seed
set.seed(1)
X <- runif(20000, min = 0, max = 2.5) # covariate
H <- rbinom(20000, 1, 0.1) # hypothesis true or false
Z <- rnorm(20000, H * X) # Z-score
.Random.seed <- save.seed
pvalue <- 1 - pnorm(Z) # pvalue
run_sim(pvalue, X, H, 1)
}
