% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_betamix.R
\name{betamix_datadriven_lfdr}
\alias{betamix_datadriven_lfdr}
\title{Local false discovery rate procedure with local fdrs estimated from the Betamix-model through the EM algorithml
from https://github.com/Huber-group-EMBL/covariate-powered-cross-weighted-multiple-testing/blob/40d485f972fd2c03b371b98d3cad3d7f1a48520c/IHWStatsPaper/R/betamix.R#L76}
\usage{
betamix_datadriven_lfdr(Ps, Xs, alpha, formula_rhs = "~.", maxiter = 200, ...)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Xs}{Data frame with features}

\item{alpha}{Significance level at which to apply method}

\item{formula_rhs}{Formula defining the RHS in the fitted GLMs, defaults to ~X1+X2 (used in simulations herein).}

\item{maxiter}{Total number of iterations to run the EM algorithm}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Local false discovery rate procedure with local fdrs estimated from the Betamix-model through the EM algorithml
from https://github.com/Huber-group-EMBL/covariate-powered-cross-weighted-multiple-testing/blob/40d485f972fd2c03b371b98d3cad3d7f1a48520c/IHWStatsPaper/R/betamix.R#L76
}
