% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_data_handling.R
\name{create_dataframe}
\alias{create_dataframe}
\title{This function iteratively creates a data.frame from a list of vectors.}
\usage{
create_dataframe(vec_list)
}
\arguments{
\item{vec_list}{List of vectors.}
}
\value{
Data frame resulting from the operation.
}
\description{
It iterates over all vectors in the list, applies expand.grid to each vector 
with the first element of all other vectors, row binds all the data frames 
and then returns the resulting data frame.
}
\examples{
vec1 <- c(1,2,3)
vec2 <- c(4,5,6)
vec3 <- c(7,8,9)
vec_list <- list(vec1, vec2, vec3)
create_dataframe(vec_list)
}
