% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{eval_sim_parallel}
\alias{eval_sim_parallel}
\title{Evaluate Simulation Parallelly}
\usage{
eval_sim_parallel(
  simulation_list,
  alpha = 0.1,
  methods = c("IHW-quantile", "IHW-forest", "IHW-forest-drop-inbag", "BH", "AdaPT",
    "Boca-Leek", "Clfdr-EM"),
  forest_par = NULL,
  null_proportion = TRUE,
  folds = NULL,
  context_information = NULL
)
}
\arguments{
\item{simulation_list}{A list of simulations.}

\item{alpha}{The alpha level used in the analysis.}

\item{methods}{Vector of methods to be used for multiple testing procedures.}

\item{forest_par}{Parameters for the forest method.}

\item{null_proportion}{Logical, if TRUE the proportion of null hypotheses is estimated.}

\item{folds}{The number of folds for cross-validation.}

\item{context_information}{Vector of strings, corresponding to elements in sim_i.}
}
\value{
A list of results from running the simulations.
}
\description{
This function is designed to evaluate a list of simulations in parallel using a user-specified number of cores.
The simulations are run for each method given in the `methods` argument.
}
\examples{
sim_list <- list() # Add simulations to this list
results <- eval_sim_parallel(simulation_list = sim_list, context_information = c("info1", "info2"))
}
