% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_flexible_rule.R
\name{flexible_prop_alt_sim}
\alias{flexible_prop_alt_sim}
\title{Simulate alternative proportions flexibly}
\usage{
flexible_prop_alt_sim(
  m,
  r,
  dimensions,
  prop_alt_function_creator = discrete_prop_alt_creator,
  prop_alt_function_name = "discrete_prop_alt_creator",
  additional_arguments_prop_alt_function_creator = NULL
)
}
\arguments{
\item{m}{A vector specifying the number of observations per simulation}

\item{r}{An integer specifying the number of replications for the simulation}

\item{dimensions}{A vector specifying the dimensionality of the covariates in the simulations}

\item{prop_alt_function_creator}{A function that creates alternative proportion functions, default is 'discrete_prop_alt_creator'}

\item{prop_alt_function_name}{A string naming the proportion alternative function creator, default is "discrete_prop_alt_creator"}

\item{additional_arguments_prop_alt_function_creator}{A dataframe that contains additional arguments needed for the prop_alt_function_creator function}
}
\value{
A list containing the simulation results for each combination of m, dimension, and seed, including additional arguments if provided
}
\description{
This function is designed to perform multiple simulations 
to calculate the proportion of alternative cases given various conditions.
}
\examples{
flexible_prop_alt_sim(m = c(100,200), r = 2, dimensions = c(2,3))

}
