% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_wrapper.R
\name{ihw_forest_wrapper}
\alias{ihw_forest_wrapper}
\title{Wrapper for the IHW-Forest procedure}
\usage{
ihw_forest_wrapper(
  Ps,
  Xs,
  alpha,
  forest_par,
  null_proportion = T,
  per_covariate_bins = 5,
  drop_inbag = FALSE
)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Xs}{Vector or matrix of covariates}

\item{alpha}{Significance level at which to apply method}

\item{forest_par}{TODO}

\item{drop_inbag}{do not use hypotheses, which have already been used for the 
construction of the partition for the construction of the weights as well}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Wrapper for the IHW-Forest procedure
}
